refactor: files 리팩토링
