+++ src/components/Button/TypeSelectButton.jsx
@@ -1,7 +1,15 @@
 import { Button } from '@chakra-ui/react';
 import React from 'react';
 
-const TypeSelectButton = (fileType, setFileType, getFilesByType, text, type) => {
+const TypeSelectButton = ({
+  fileType,
+  setFileType,
+  getFilesByType,
+  text,
+  type,
+  workSpaceId,
+  setFileList,
+}) => {
   return (
     <Button
       borderRadius="8px"
@@ -11,10 +19,12 @@ const TypeSelectButton = (fileType, setFileType, getFilesByType, text, type) => 
       bgColor={fileType === type ? '#575DFB' : '#FFFFFF'}
       color={fileType === type ? 'white' : 'black'}
       _hover={fileType === type ? { bgColor: '#575DFB' } : undefined}
-      border={fileType === type ? '1px solid #575DFB' : '1px solid #898989'}
+      border={
+        fileType.type === type ? '1px solid #575DFB' : '1px solid #898989'
+      }
       onClick={() => {
-        setFileType({ fileType });
-        getFilesByType({ fileType });
+        setFileType(type);
+        getFilesByType({ type, workSpaceId, setFileList });
       }}
     >
       {text}
